set default journal mode to TRUNCATE, not PERSIST
PERSIST causes data to linger in the journal beyond transactions, so
it interacts poorly with SECURE_DELETE. It's also not necessarily
faster, especially since it makes the page cache and SSD cache (if there
is one) colder by holding on to blocks that are no longer used.

The default chosen by upstream is DELETE, but TRUNCATE is a bit lighter
since it doesn't need to create/delete the file.

Change-Id: If790271c253a840df241980f538fc0d844ba7857
Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>

Conflicts:
	core/res/res/values/config.xml